postShort - add image hover
